refactor(sandbox): split adapters + dedup sleep/cloudflare-retry helper
