+++ packages/sandbox/src/internal/shared/utils.ts
@@ -1,0 +1,3 @@
+export async function sleep(ms: number): Promise<void> {
+  await new Promise(resolve => setTimeout(resolve, ms))
+}
