+++ packages/sandbox/src/sandbox/adapters/_shared.ts
@@ -1,15 +1,14 @@
 import type { SandboxWaitForPortOptions } from '../types/common'
 import { hasProtocol } from 'ufo'
+import { sleep } from '../../internal/shared/utils'
 import { SandboxError } from '../errors'
+
+export { sleep }
 
 export function normalizeLogPattern(pattern: string | RegExp): RegExp {
   return typeof pattern === 'string'
     ? new RegExp(pattern)
     : new RegExp(pattern.source, pattern.flags.replace(/g/g, ''))
-}
-
-export async function sleep(ms: number): Promise<void> {
-  await new Promise(resolve => setTimeout(resolve, ms))
 }
 
 function resolvePortProbeUrl(
