+++ packages/sandbox/src/sandbox/adapters/cloudflare.ts
@@ -1,15 +1,18 @@
-import type { CloudflareSandboxNamespace, CloudflareSandboxSession, CloudflareSandboxSessionOptions, SandboxCodeContext, SandboxCodeContextOptions, SandboxCodeExecutionResult, SandboxExposedPort, SandboxExposePortOptions, SandboxGitCheckoutOptions, SandboxGitCheckoutResult, SandboxMountBucketOptions, SandboxRunCodeOptions } from '../types/cloudflare'
-import type { CloudflareSandboxStub, SandboxCapabilities, SandboxExecOptions, SandboxExecResult, SandboxFileEntry, SandboxListFilesOptions, SandboxProcess, SandboxProcessOptions, SandboxWaitForPortOptions } from '../types/common'
-import { CLOUDFLARE_RETRIABLE_STARTUP_ERROR_RE, CLOUDFLARE_SANDBOX_RETRY_DELAYS_MS } from '../../internal/shared/cloudflare-retry'
 import { NotSupportedError, SandboxError } from '../errors'
 import { shellQuote } from '../utils'
-import { normalizeLogPattern, sleep } from './_shared'
 import { BaseSandboxAdapter } from './base'
+import { CloudflareProcessHandle, type CloudflareProcessHandleCompat } from './cloudflare/process'
+import {
+  CLOUDFLARE_CONTROL_PLANE_TIMEOUT_MS,
+  CLOUDFLARE_READ_FILE_TIMEOUT_MS,
+  CLOUDFLARE_STOP_TIMEOUT_MS,
+  resolveExecRequestTimeout,
+  withCloudflareDeadline,
+  withCloudflareTransportRetry,
+} from './cloudflare/transport'
 
-const CLOUDFLARE_CONTROL_PLANE_TIMEOUT_MS = 15_000
-const CLOUDFLARE_READ_FILE_TIMEOUT_MS = 15_000
-const CLOUDFLARE_STOP_TIMEOUT_MS = 10_000
-const CLOUDFLARE_EXEC_REQUEST_TIMEOUT_MS = 180_000
+import type { CloudflareSandboxNamespace, CloudflareSandboxSession, CloudflareSandboxSessionOptions, SandboxCodeContext, SandboxCodeContextOptions, SandboxCodeExecutionResult, SandboxExposedPort, SandboxExposePortOptions, SandboxGitCheckoutOptions, SandboxGitCheckoutResult, SandboxMountBucketOptions, SandboxRunCodeOptions } from '../types/cloudflare'
+import type { CloudflareSandboxStub, SandboxCapabilities, SandboxExecOptions, SandboxExecResult, SandboxFileEntry, SandboxListFilesOptions, SandboxProcess, SandboxProcessOptions } from '../types/common'
 
 type CloudflareExtendedStub = CloudflareSandboxStub & {
   gitCheckout?: (url: string, opts?: SandboxGitCheckoutOptions) => Promise<SandboxGitCheckoutResult>
@@ -27,150 +30,6 @@ type CloudflareExtendedStub = CloudflareSandboxStub & {
   unmountBucket?: (path: string) => Promise<void>
   setEnvVars?: (vars: Record<string, string | undefined>) => Promise<void>
   wsConnect?: (request: Request, port: number) => Promise<Response>
-}
-
-type CloudflareProcessHandleCompat = {
-  id: string
-  command: string
-  kill: (signal?: string) => Promise<void>
-  getLogs: () => Promise<{ stdout: string, stderr: string }>
-  waitForExit: (timeout?: number) => Promise<{ exitCode: number }>
-  waitForLog: (pattern: string | RegExp, timeout?: number) => Promise<{ line: string }>
-  waitForPort: (port: number, opts?: { timeout?: number, hostname?: string }) => Promise<void>
-}
-
-function createCloudflareTransportError(operation: string, error: unknown) {
-  const message = error instanceof Error ? error.message : String(error)
-  return new SandboxError(message, {
-    cause: error,
-    code: 'SANDBOX_TRANSPORT_ERROR',
-    details: { operation },
-    provider: 'cloudflare',
-  })
-}
-
-function isRetriableCloudflareTransportError(error: unknown) {
-  const sandboxError = error instanceof SandboxError ? error : undefined
-  const message = error instanceof Error ? error.message : String(error)
-  if (sandboxError?.code === 'TIMEOUT')
-    return true
-  return CLOUDFLARE_RETRIABLE_STARTUP_ERROR_RE.test(message)
-}
-
-async function withCloudflareDeadline<T>(operation: string, timeoutMs: number, run: () => Promise<T>) {
-  let timeoutId: ReturnType<typeof setTimeout> | undefined
-
-  try {
-    return await Promise.race([
-      run(),
-      new Promise<never>((_, reject) => {
-        timeoutId = setTimeout(() => {
-          reject(new SandboxError(`Cloudflare sandbox ${operation} timed out after ${timeoutMs}ms.`, {
-            code: 'TIMEOUT',
-            details: { operation, timeout: timeoutMs },
-            provider: 'cloudflare',
-          }))
-        }, timeoutMs)
-      }),
-    ])
-  }
-  finally {
-    if (timeoutId)
-      clearTimeout(timeoutId)
-  }
-}
-
-async function withCloudflareTransportRetry<T>(operation: string, run: () => Promise<T>) {
-  const attempts = CLOUDFLARE_SANDBOX_RETRY_DELAYS_MS.length + 1
-
-  for (let attempt = 0; attempt < attempts; attempt++) {
-    try {
-      return await run()
-    }
-    catch (error) {
-      const sandboxError = error instanceof SandboxError
-        ? error
-        : createCloudflareTransportError(operation, error)
-
-      const shouldRetry = attempt < CLOUDFLARE_SANDBOX_RETRY_DELAYS_MS.length
-        && isRetriableCloudflareTransportError(sandboxError)
-
-      if (!shouldRetry)
-        throw sandboxError
-
-      await sleep(CLOUDFLARE_SANDBOX_RETRY_DELAYS_MS[attempt])
-    }
-  }
-
-  throw new SandboxError(`Cloudflare sandbox ${operation} retries exhausted.`, {
-    code: 'SANDBOX_TRANSPORT_ERROR',
-    details: { operation },
-    provider: 'cloudflare',
-  })
-}
-
-function resolveExecRequestTimeout(timeout?: number) {
-  if (typeof timeout === 'number' && timeout > 0)
-    return Math.min(timeout, CLOUDFLARE_EXEC_REQUEST_TIMEOUT_MS)
-
-  return CLOUDFLARE_EXEC_REQUEST_TIMEOUT_MS
-}
-
-class CloudflareProcessHandle implements SandboxProcess {
-  readonly id: string
-  readonly command: string
-  private processInfo: {
-    kill: (signal?: string) => Promise<void>
-    getLogs: () => Promise<{ stdout: string, stderr: string }>
-    waitForExit: (timeout?: number) => Promise<{ exitCode: number }>
-    waitForLog: (pattern: string | RegExp, timeout?: number) => Promise<{ line: string }>
-    waitForPort: (port: number, opts?: { timeout?: number, hostname?: string }) => Promise<void>
-  }
-
-  constructor(id: string, command: string, processInfo: CloudflareProcessHandle['processInfo']) {
-    this.id = id
-    this.command = command
-    this.processInfo = processInfo
-  }
-
-  async kill(signal?: string): Promise<void> {
-    await this.processInfo.kill(signal)
-  }
-
-  async logs(): Promise<{ stdout: string, stderr: string }> {
-    return this.processInfo.getLogs()
-  }
-
-  async wait(timeout?: number): Promise<{ exitCode: number }> {
-    return this.processInfo.waitForExit(timeout)
-  }
-
-  async waitForLog(pattern: string | RegExp, timeout = 30_000): Promise<{ line: string }> {
-    try {
-      return await this.processInfo.waitForLog(pattern, timeout)
-    }
-    catch (error) {
-      const message = error instanceof Error ? error.message : String(error)
-      if (!/ProcessExitedBeforeReadyError|before becoming ready|exited with code/i.test(message))
-        throw error
-    }
-
-    const startTime = Date.now()
-    const regex = normalizeLogPattern(pattern)
-    while (Date.now() - startTime < timeout) {
-      const { stdout, stderr } = await this.logs()
-      for (const line of `${stdout}${stderr}`.split('\n')) {
-        if (regex.test(line))
-          return { line }
-      }
-      await sleep(100)
-    }
-    throw new SandboxError(`Timeout waiting for log pattern: ${pattern}`, 'TIMEOUT')
-  }
-
-  async waitForPort(port: number, opts?: SandboxWaitForPortOptions): Promise<void> {
-    await this.processInfo.waitForPort(port, opts)
-  }
 }
 
 class CloudflareNamespaceImpl implements CloudflareSandboxNamespace {
